Move ToUrlSlug to Parser, fix issue with trakt lists that have non-alphanumeric characters
